# Directly store values in GroupKey and avoid ser/de