--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/DictionaryBasedGroupKeyGenerator.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/DictionaryBasedGroupKeyGenerator.java
@@ -27,7 +27,6 @@ import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
 import it.unimi.dsi.fastutil.objects.ObjectIterator;
 import java.util.Arrays;
 import java.util.Iterator;
-import java.util.NoSuchElementException;
 import org.apache.pinot.core.common.BlockValSet;
 import org.apache.pinot.core.operator.blocks.TransformBlock;
 import org.apache.pinot.core.operator.transform.TransformOperator;
@@ -189,11 +188,16 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
   }
 
   @Override
-  public Iterator<GroupKey> getUniqueGroupKeys() {
-    return _rawKeyHolder.iterator();
+  public Iterator<GroupKey> getGroupKeys() {
+    return _rawKeyHolder.getGroupKeys();
   }
 
-  private interface RawKeyHolder extends Iterable<GroupKey> {
+  @Override
+  public Iterator<StringGroupKey> getStringGroupKeys() {
+    return _rawKeyHolder.getStringGroupKeys();
+  }
+
+  private interface RawKeyHolder {
 
     /**
      * Process a block of documents for all single-valued group-by columns case.
@@ -217,6 +221,16 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
      * @return Upper bound of group id inside the holder
      */
     int getGroupIdUpperBound();
+
+    /**
+     * Returns an iterator of {@link GroupKey}. Use this interface to iterate through all the group keys.
+     */
+    Iterator<GroupKey> getGroupKeys();
+
+    /**
+     * Returns an iterator of {@link StringGroupKey}. Use this interface to iterate through all the group keys.
+     */
+    Iterator<StringGroupKey> getStringGroupKeys();
   }
 
   private class ArrayBasedHolder implements RawKeyHolder {
@@ -252,7 +266,7 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
     }
 
     @Override
-    public Iterator<GroupKey> iterator() {
+    public Iterator<GroupKey> getGroupKeys() {
       return new Iterator<GroupKey>() {
         private int _currentGroupId;
         private final GroupKey _groupKey = new GroupKey();
@@ -267,11 +281,37 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
 
         @Override
         public GroupKey next() {
-          if (!hasNext()) {
-            throw new NoSuchElementException();
-          }
           _groupKey._groupId = _currentGroupId;
-          _groupKey._stringKey = getGroupKey(_currentGroupId);
+          _groupKey._keys = getKeys(_currentGroupId);
+          _currentGroupId++;
+          return _groupKey;
+        }
+
+        @Override
+        public void remove() {
+          throw new UnsupportedOperationException();
+        }
+      };
+    }
+
+    @Override
+    public Iterator<StringGroupKey> getStringGroupKeys() {
+      return new Iterator<StringGroupKey>() {
+        private int _currentGroupId;
+        private final StringGroupKey _groupKey = new StringGroupKey();
+
+        @Override
+        public boolean hasNext() {
+          while (_currentGroupId < _globalGroupIdUpperBound && !_flags[_currentGroupId]) {
+            _currentGroupId++;
+          }
+          return _currentGroupId < _globalGroupIdUpperBound;
+        }
+
+        @Override
+        public StringGroupKey next() {
+          _groupKey._groupId = _currentGroupId;
+          _groupKey._stringKey = getStringKey(_currentGroupId);
           _currentGroupId++;
           return _groupKey;
         }
@@ -320,10 +360,10 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
     }
 
     @Override
-    public Iterator<GroupKey> iterator() {
+    public Iterator<GroupKey> getGroupKeys() {
       return new Iterator<GroupKey>() {
-        private final GroupKey _groupKey = new GroupKey();
         private final Iterator<IntGroupIdMap.Entry> _iterator = _groupIdMap.iterator();
+        private final GroupKey _groupKey = new GroupKey();
 
         @Override
         public boolean hasNext() {
@@ -334,7 +374,33 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
         public GroupKey next() {
           IntGroupIdMap.Entry entry = _iterator.next();
           _groupKey._groupId = entry._groupId;
-          _groupKey._stringKey = getGroupKey(entry._rawKey);
+          _groupKey._keys = getKeys(entry._rawKey);
+          return _groupKey;
+        }
+
+        @Override
+        public void remove() {
+          throw new UnsupportedOperationException();
+        }
+      };
+    }
+
+    @Override
+    public Iterator<StringGroupKey> getStringGroupKeys() {
+      return new Iterator<StringGroupKey>() {
+        private final Iterator<IntGroupIdMap.Entry> _iterator = _groupIdMap.iterator();
+        private final StringGroupKey _groupKey = new StringGroupKey();
+
+        @Override
+        public boolean hasNext() {
+          return _iterator.hasNext();
+        }
+
+        @Override
+        public StringGroupKey next() {
+          IntGroupIdMap.Entry entry = _iterator.next();
+          _groupKey._groupId = entry._groupId;
+          _groupKey._stringKey = getStringKey(entry._rawKey);
           return _groupKey;
         }
 
@@ -425,12 +491,27 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
   }
 
   /**
-   * Helper method to get group key from raw key.
-   *
-   * @param rawKey Integer raw key
-   * @return String group key
+   * Helper method to get the keys from the raw key.
    */
-  private String getGroupKey(int rawKey) {
+  private Object[] getKeys(int rawKey) {
+    // Specialize single group-by column case
+    if (_numGroupByExpressions == 1) {
+      return new Object[]{_dictionaries[0].getInternal(rawKey)};
+    } else {
+      Object[] groupKeys = new Object[_numGroupByExpressions];
+      for (int i = 0; i < _numGroupByExpressions; i++) {
+        int cardinality = _cardinalities[i];
+        groupKeys[i] = _dictionaries[i].getInternal(rawKey % cardinality);
+        rawKey /= cardinality;
+      }
+      return groupKeys;
+    }
+  }
+
+  /**
+   * Helper method to get the string key from the raw key.
+   */
+  private String getStringKey(int rawKey) {
     // Specialize single group-by column case
     if (_numGroupByExpressions == 1) {
       return _dictionaries[0].getStringValue(rawKey);
@@ -494,7 +575,7 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
     }
 
     @Override
-    public Iterator<GroupKey> iterator() {
+    public Iterator<GroupKey> getGroupKeys() {
       return new Iterator<GroupKey>() {
         private final ObjectIterator<Long2IntMap.Entry> _iterator = _groupIdMap.long2IntEntrySet().fastIterator();
         private final GroupKey _groupKey = new GroupKey();
@@ -508,7 +589,33 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
         public GroupKey next() {
           Long2IntMap.Entry entry = _iterator.next();
           _groupKey._groupId = entry.getIntValue();
-          _groupKey._stringKey = getGroupKey(entry.getLongKey());
+          _groupKey._keys = getKeys(entry.getLongKey());
+          return _groupKey;
+        }
+
+        @Override
+        public void remove() {
+          throw new UnsupportedOperationException();
+        }
+      };
+    }
+
+    @Override
+    public Iterator<StringGroupKey> getStringGroupKeys() {
+      return new Iterator<StringGroupKey>() {
+        private final ObjectIterator<Long2IntMap.Entry> _iterator = _groupIdMap.long2IntEntrySet().fastIterator();
+        private final StringGroupKey _groupKey = new StringGroupKey();
+
+        @Override
+        public boolean hasNext() {
+          return _iterator.hasNext();
+        }
+
+        @Override
+        public StringGroupKey next() {
+          Long2IntMap.Entry entry = _iterator.next();
+          _groupKey._groupId = entry.getIntValue();
+          _groupKey._stringKey = getStringKey(entry.getLongKey());
           return _groupKey;
         }
 
@@ -594,12 +701,22 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
   }
 
   /**
-   * Helper method to get group key from raw key.
-   *
-   * @param rawKey Long raw key
-   * @return String group key
+   * Helper method to get the keys from the raw key.
    */
-  private String getGroupKey(long rawKey) {
+  private Object[] getKeys(long rawKey) {
+    Object[] groupKeys = new Object[_numGroupByExpressions];
+    for (int i = 0; i < _numGroupByExpressions; i++) {
+      int cardinality = _cardinalities[i];
+      groupKeys[i] = _dictionaries[i].getInternal((int) (rawKey % cardinality));
+      rawKey /= cardinality;
+    }
+    return groupKeys;
+  }
+
+  /**
+   * Helper method to get the string key from the raw key.
+   */
+  private String getStringKey(long rawKey) {
     int cardinality = _cardinalities[0];
     StringBuilder groupKeyBuilder = new StringBuilder(_dictionaries[0].getStringValue((int) (rawKey % cardinality)));
     rawKey /= cardinality;
@@ -658,7 +775,7 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
     }
 
     @Override
-    public Iterator<GroupKey> iterator() {
+    public Iterator<GroupKey> getGroupKeys() {
       return new Iterator<GroupKey>() {
         private final ObjectIterator<Object2IntMap.Entry<IntArray>> _iterator =
             _groupIdMap.object2IntEntrySet().fastIterator();
@@ -673,7 +790,34 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
         public GroupKey next() {
           Object2IntMap.Entry<IntArray> entry = _iterator.next();
           _groupKey._groupId = entry.getIntValue();
-          _groupKey._stringKey = getGroupKey(entry.getKey());
+          _groupKey._keys = getKeys(entry.getKey());
+          return _groupKey;
+        }
+
+        @Override
+        public void remove() {
+          throw new UnsupportedOperationException();
+        }
+      };
+    }
+
+    @Override
+    public Iterator<StringGroupKey> getStringGroupKeys() {
+      return new Iterator<StringGroupKey>() {
+        private final ObjectIterator<Object2IntMap.Entry<IntArray>> _iterator =
+            _groupIdMap.object2IntEntrySet().fastIterator();
+        private final StringGroupKey _groupKey = new StringGroupKey();
+
+        @Override
+        public boolean hasNext() {
+          return _iterator.hasNext();
+        }
+
+        @Override
+        public StringGroupKey next() {
+          Object2IntMap.Entry<IntArray> entry = _iterator.next();
+          _groupKey._groupId = entry.getIntValue();
+          _groupKey._stringKey = getStringKey(entry.getKey());
           return _groupKey;
         }
 
@@ -763,12 +907,20 @@ public class DictionaryBasedGroupKeyGenerator implements GroupKeyGenerator {
   }
 
   /**
-   * Helper method to get group key from raw key.
-   *
-   * @param rawKey IntArray raw key
-   * @return String group key
+   * Helper method to get the keys from the raw key.
    */
-  private String getGroupKey(IntArray rawKey) {
+  private Object[] getKeys(IntArray rawKey) {
+    Object[] groupKeys = new Object[_numGroupByExpressions];
+    for (int i = 0; i < _numGroupByExpressions; i++) {
+      groupKeys[i] = _dictionaries[i].getInternal(rawKey._elements[i]);
+    }
+    return groupKeys;
+  }
+
+  /**
+   * Helper method to get the string key from the raw key.
+   */
+  private String getStringKey(IntArray rawKey) {
     StringBuilder groupKeyBuilder = new StringBuilder(_dictionaries[0].getStringValue(rawKey._elements[0]));
     for (int i = 1; i < _numGroupByExpressions; i++) {
       groupKeyBuilder.append(GroupKeyGenerator.DELIMITER);