--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/BytesToIdMap.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/BytesToIdMap.java
@@ -59,4 +59,9 @@ public class BytesToIdMap extends BaseValueToIdMap {
     assert id < _idToValueMap.size();
     return _idToValueMap.get(id);
   }
+
+  @Override
+  public Object get(int id) {
+    return getBytes(id);
+  }
 }