--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/DoubleToIdMap.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/DoubleToIdMap.java
@@ -58,4 +58,9 @@ public class DoubleToIdMap extends BaseValueToIdMap {
   public String getString(int id) {
     return Double.toString(getDouble(id));
   }
+
+  @Override
+  public Object get(int id) {
+    return getDouble(id);
+  }
 }