--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/FloatToIdMap.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/FloatToIdMap.java
@@ -58,4 +58,9 @@ public class FloatToIdMap extends BaseValueToIdMap {
   public String getString(int id) {
     return Float.toString(getFloat(id));
   }
+
+  @Override
+  public Object get(int id) {
+    return getFloat(id);
+  }
 }