--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/IntToIdMap.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/IntToIdMap.java
@@ -58,4 +58,9 @@ public class IntToIdMap extends BaseValueToIdMap {
   public String getString(int id) {
     return Integer.toString(getInt(id));
   }
+
+  @Override
+  public Object get(int id) {
+    return getInt(id);
+  }
 }