--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/LongToIdMap.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/LongToIdMap.java
@@ -58,4 +58,9 @@ public class LongToIdMap extends BaseValueToIdMap {
   public String getString(int id) {
     return Long.toString(getLong(id));
   }
+
+  @Override
+  public Object get(int id) {
+    return getLong(id);
+  }
 }