--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/StringToIdMap.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/utils/StringToIdMap.java
@@ -53,4 +53,9 @@ public class StringToIdMap extends BaseValueToIdMap {
     assert id < _idToValueMap.size();
     return _idToValueMap.get(id);
   }
+
+  @Override
+  public Object get(int id) {
+    return getString(id);
+  }
 }