--- a/pinot-core/src/main/java/org/apache/pinot/core/segment/index/readers/BytesDictionary.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/segment/index/readers/BytesDictionary.java
@@ -59,6 +59,11 @@ public class BytesDictionary extends BaseImmutableDictionary {
   }
 
   @Override
+  public Object getInternal(int dictId) {
+    return new ByteArray(getBytes(dictId));
+  }
+
+  @Override
   public int getIntValue(int dictId) {
     throw new UnsupportedOperationException();
   }