… pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/groupby/NoDictionarySingleColumnGroupKeyGenerator.java
… pinot-core/src/main/java/org/apache/pinot/core/segment/index/readers/ConstantValueBytesDictionary.java
… pinot-core/src/main/java/org/apache/pinot/core/segment/index/readers/Dictionary.java
… pinot-core/src/test/java/org/apache/pinot/queries/DistinctCountBitmapQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/DistinctCountQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/DistinctCountThetaSketchQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/FSTBasedRegexpLikeQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/FastHllQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/IdSetQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/PercentileTDigestQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/QueriesTestUtils.java
… pinot-core/src/test/java/org/apache/pinot/queries/SegmentPartitionedDistinctCountQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/SerializedBytesQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/StUnionQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/TransformQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/query/aggregation/groupby/DictionaryBasedGroupKeyGeneratorTest.java
… pinot-core/src/test/java/org/apache/pinot/query/aggregation/groupby/NoDictionaryGroupKeyGeneratorTest.java
… pinot-core/src/test/java/org/apache/pinot/query/aggregation/groupby/StringGroupKeyTest.java